+++ app/controllers/dashboard_controller.rb
@@ -5,6 +5,8 @@ class DashboardController < ApplicationController
 
   def show
     @user = Current.user
+    @games = @user.game_events.includes(:game_system, :game_participations, :players)
+                  .order(played_at: :desc)
   end
 
   private