+++ app/javascript/controllers/index.js
@@ -2,3 +2,6 @@
 import { application } from "controllers/application"
 import { eagerLoadControllersFrom } from "@hotwired/stimulus-loading"
 eagerLoadControllersFrom("controllers", application)
+
+import GameFormController from "controllers/game_form_controller"
+application.register("game-form", GameFormController)