+++ app/models/game/event.rb
@@ -8,13 +8,49 @@ module Game
              foreign_key: 'game_event_id',
              inverse_of: :game_event,
              dependent: :destroy
-    has_many :players, through: :participations, source: :user
+    accepts_nested_attributes_for :game_participations
+    has_many :players, through: :game_participations, source: :user
 
     validates :played_at, presence: true
-    # validate :must_have_at_least_two_players
+    validate :must_have_exactly_two_players
+    validate :players_must_be_distinct
+    validate :both_scores_present
 
-    # def must_have_at_least_two_players
-    #   errors.add(:players, 'must have at least two players') if players.size < 2
-    # end
+    def winner_user
+      participations = game_participations.to_a
+      return nil unless participations.size == 2
+
+      a, b = participations
+      return nil if a.score == b.score
+
+      a.score > b.score ? a.user : b.user
+    end
+
+    private
+
+    def must_have_exactly_two_players
+      return if game_participations.reject(&:marked_for_destruction?).size == 2
+
+      errors.add(:players, I18n.t('games.errors.exactly_two_players'))
+    end
+
+    def players_must_be_distinct
+      participations = game_participations.reject(&:marked_for_destruction?)
+      return unless participations.size == 2
+
+      user_ids = participations.map(&:user_id)
+      return unless user_ids.all?(&:present?)
+
+      errors.add(:players, I18n.t('games.errors.exactly_two_players')) if user_ids.uniq.size != 2
+    end
+
+    def both_scores_present
+      participations = game_participations.reject(&:marked_for_destruction?)
+      return unless participations.size == 2
+
+      return unless participations.any? { |p| p.score.blank? }
+
+      errors.add(:players, I18n.t('games.errors.both_scores_required'))
+    end
   end
 end