+++ app/models/game/participation.rb
@@ -5,7 +5,7 @@ module Game
     belongs_to :game_event, class_name: 'Game::Event'
     belongs_to :user
 
-    validates :result, presence: true
+    validates :score, presence: true
     validates :user_id, uniqueness: { scope: :game_event_id }
   end
 end