+++ test/controllers/pages_controller_test.rb
@@ -9,25 +9,15 @@ class PagesControllerTest < ActionDispatch::IntegrationTest
     assert_select 'h1', 'Welcome to Uniladder'
   end
 
-  test 'should get home when logged in' do
+  test 'should redirect to dashboard when logged in' do
     user = users(:player_one)
     post session_path, params: { email_address: user.email_address, password: 'password' }
 
     get root_path
-    assert_response :success
-    assert_select 'h1', 'Welcome to Uniladder'
+    assert_redirected_to dashboard_path(locale: I18n.locale)
   end
 
   test 'should get home in French' do
-    get root_path(locale: :fr)
-    assert_response :success
-    assert_select 'h1', 'Bienvenue sur Uniladder'
-  end
-
-  test 'should get home in French when logged in' do
-    user = users(:player_one)
-    post session_path, params: { email_address: user.email_address, password: 'password' }
-
     get root_path(locale: :fr)
     assert_response :success
     assert_select 'h1', 'Bienvenue sur Uniladder'